chore: name test files after source files, not behaviors
